+++ pkg/resources/hpas/hpa.go
@@ -203,8 +203,8 @@ func Resolve(src orktypes.HPATemplateSource, ownerName string, reg orktypes.Prof
 		spec.Behavior = &b
 	}
 
-	for _, l := range src.Labels {
-		spec.Labels[l.Key] = l.Value
+	for k, v := range src.Labels {
+		spec.Labels[k] = v
 	}
 
 	// System labels