+++ pkg/resources/limitranges/limitrange.go
@@ -231,8 +231,8 @@ func Resolve(src orktypes.LimitRangeTemplateSource, ownerName string, reg orktyp
 		Sleep:          src.Sleep,
 	}
 
-	for _, l := range src.Labels {
-		spec.Labels[l.Key] = l.Value
+	for k, v := range src.Labels {
+		spec.Labels[k] = v
 	}
 	spec.Labels[labels.ManagedKey] = labels.ManagedValue
 	spec.Labels[labels.OrkestraOwner] = ownerName